+++ src/BlockBuilder/FieldType/Type/FilesFromFolder/FilesFromFolderFieldTypeDto.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types=1);
+
+namespace BlockBuilder\FieldType\Type\FilesFromFolder;
+
+use BlockBuilder\FieldType\Enum\FieldTypeEnum;
+use BlockBuilder\FieldType\FieldTypeDtoInterface;
+
+final readonly class FilesFromFolderFieldTypeDto implements FieldTypeDtoInterface
+{
+    public function __construct(
+        public FieldTypeEnum $fieldType,
+        public string $label,
+        public string $handle,
+        public bool $required,
+        public ?string $helpText,
+        public string $fileOrder,
+    ) {
+    }
+}